Revert "src/bin/pg_ctl/pg_ctl.c: add patch for pg_ctl race with ntpd"
This temporarily reverts commit e662d1862c833d81da9db1c95e36a12761d6e61c.
to prevent merge conflict during import.
Because of the changed codebase this patch would be splitted into
the original patch and merge commit. We would like to avoid confusion
so we will add the patch again to the changed code. It also makes it
easier to port to other repos.